+++ app/edit_selection.c
@@ -447,7 +447,7 @@ edit_selection_motion (Tool           *tool,
     }
   else /* show real world units */
     {
-      float unit_factor = gimp_unit_get_factor (gdisp->gimage->unit);
+      gdouble unit_factor = gimp_unit_get_factor (gdisp->gimage->unit);
 
       g_snprintf (offset, STATUSBAR_SIZE, gdisp->cursor_format_str,
 		  _("Move: "),
